add creation of TM-BOM scope
