+++ td.vue/src/service/migration/tmBom/detail.js
@@ -1,11 +1,5 @@
 import diagrams from './diagrams/diagrams';
 
-const convert = (model) => {
-    return {
-	    diagrams: diagrams.convert(model)
-    };
-};
-
 const merge = (model, version) => {
     let allDiagrams = diagrams.merge(model, version);
     
@@ -19,6 +13,5 @@ const merge = (model, version) => {
 };
 
 export default {
-    convert,
     merge
 };
