+++ td.vue/tests/unit/service/migration/tmBom/scope.spec.js
@@ -1,0 +1,49 @@
+import scope from '@/service/migration/tmBom/scope';
+import tdModel from './td-test-model';
+
+jest.mock('@/service/migration/tmBom/diagrams/assumptions');
+
+describe('service/migration/tmBom/scope.js', () => {
+
+    describe('convert TM-BOM', () => {
+        let testScope;
+
+	    describe('recreates TM-BOM scope', () => {
+	        beforeEach(() => {
+	            testScope = scope.convert(tdModel);
+	        });
+	
+	        it('populates the scope values', () => {
+	            expect(testScope.title).toBe(tdModel.summary.title);
+	            expect(testScope.description).toBe(tdModel.summary.description);
+	        });
+	
+	        it('reinstates the compatibility values to scope', () => {
+	            expect(testScope.business_criticality).toBe(tdModel.summary.compatibility.business_criticality);
+	            expect(testScope.data_sensitivity).toBe(tdModel.summary.compatibility.data_sensitivity);
+	            expect(testScope.exposure).toBe(tdModel.summary.compatibility.exposure);
+	            expect(testScope.tier).toBe(tdModel.summary.compatibility.tier);
+	        });
+	    });
+
+	    describe('provides TM-BOM scope defaults', () => {
+	        it('creates default scope values', () => {
+                let noDescription = JSON.parse(JSON.stringify(tdModel));
+                delete noDescription.summary.description;
+                testScope = scope.convert(noDescription);
+	            expect(testScope.description.length).toBeGreaterThan(0);
+	        });
+
+	        it('provides scope compatibility defaults', () => {
+                let noCompatibility = JSON.parse(JSON.stringify(tdModel));
+                delete noCompatibility.summary.compatibility;
+                testScope = scope.convert(noCompatibility);
+                expect(testScope.business_criticality).not.toBeNull();
+                expect(testScope.data_sensitivity).not.toBeNull();
+                expect(testScope.exposure).not.toBeNull();
+                expect(testScope.tier).not.toBeNull();
+	        });
+	    });
+    });
+
+});
